EditSection refactoring
