+++ src/common/hooks/useSearchNavigationState.ts
@@ -8,7 +8,7 @@ import state from '@state';
 export const useSearchNavigationState = () => {
   const [searchParams] = useSearchParams();
   const querySearchParam = searchParams.get(SearchQueryParams.Query);
-  const searchBySearchParam = searchParams.get(SearchQueryParams.SearchBy);
+  const searchBySearchParam = searchParams.get(SearchQueryParams.SearchBy) as SearchIdentifiers;
   const offsetSearchParam = searchParams.get(SearchQueryParams.Offset);
 
   const setNavigationState = useSetRecoilState(state.search.navigationState);
@@ -16,7 +16,7 @@ export const useSearchNavigationState = () => {
   useEffect(() => {
     const generatedState = generateSearchParamsState(
       querySearchParam,
-      searchBySearchParam as SearchIdentifiers,
+      searchBySearchParam,
       offsetSearchParam as unknown as number,
     );
